[cita chain] seperated errors
