+++ cita-chain/types/src/errors/mod.rs
@@ -1,3 +1,13 @@
+mod block;
 mod error;
+mod execution;
+mod import;
+mod transaction;
+mod util;
 
-pub use error::*;
+pub use self::util::UtilError;
+pub use block::BlockError;
+pub use error::Error;
+pub use execution::ExecutionError;
+pub use import::ImportError;
+pub use transaction::TransactionError;
